Refactor exceptions

+++ Tournament.API/Tournament.API.Domain/Participants/ParticipantManager.cs
@@ -3,6 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
+using Tournament.API.Domain.Participants.Exceptions;
 
 namespace Tournament.API.Domain.Participants
 {
@@ -23,14 +24,21 @@ namespace Tournament.API.Domain.Participants
 
         public async Task DeleteAsync(int id)
         {
-            // TODO: check for existence
+           if(!await _repository.AnyAsync(x => x.Id == id))
+            {
+                throw new ParticipantNotExistsException(id);
+            }
+
             await _repository.DeleteAsync(id);  
         }
 
         public async Task<Participant> UpdateAsync(int id, string name, string lastName)
         {
             var participant = await _repository.GetAsync(id);
-            // TODO: Check for null
+            if(participant == null)
+            {
+                throw new ParticipantNotExistsException(id);
+            }
             participant.SetName(name);
             participant.SetLastName(lastName);
             return participant;
